Remove redundant icon composition paths in stack details
