+++ Mods/vcmi/Content/config/translations/czech.json
@@ -1038,15 +1038,15 @@
 	"vcmi.stackExperience.table.no": "Ne",
 	"vcmi.stackExperience.table.shots": "Střely",
 		"vcmi.stackExperience.table.speed": "Rychlost",
-		"vcmi.stackExperience.desc.experience": "Aktuální hodnota zkušeností oddílu, podle které se určuje hodnost a všechny bonusy ze zkušeností oddílu.",
-		"vcmi.stackExperience.desc.minDamage": "Bonus k minimálnímu poškození udělený hodností zkušeností oddílu.",
-		"vcmi.stackExperience.desc.maxDamage": "Bonus k maximálnímu poškození udělený hodností zkušeností oddílu.",
-		"vcmi.stackExperience.desc.health": "Bonus ke zdraví udělený hodností zkušeností oddílu.",
-		"vcmi.stackExperience.desc.speed": "Bonus k rychlosti udělený hodností zkušeností oddílu.",
-		"vcmi.stackExperience.desc.attack": "Bonus k útoku udělený hodností zkušeností oddílu.",
-		"vcmi.stackExperience.desc.defense": "Bonus k obraně udělený hodností zkušeností oddílu.",
-		"vcmi.stackExperience.desc.shots": "Bonus ke střelám udělený hodností zkušeností oddílu.",
-		"vcmi.stackExperience.desc.casts": "Bonus k počtu seslání udělený hodností zkušeností oddílu.",
+		"vcmi.stackExperience.desc.experience": "{Zkušenosti}\nPrahová hodnota zkušeností pro tuto hodnost je ${val}.",
+		"vcmi.stackExperience.desc.minDamage": "{Min. poškození}\nZvyšuje minimální poškození o ${val}.",
+		"vcmi.stackExperience.desc.maxDamage": "{Max. poškození}\nZvyšuje maximální poškození o ${val}.",
+		"vcmi.stackExperience.desc.health": "{Zdraví}\nZvyšuje zdraví oddílu o ${val}.",
+		"vcmi.stackExperience.desc.speed": "{Rychlost}\nZvyšuje rychlost o ${val}.",
+		"vcmi.stackExperience.desc.attack": "{Útok}\nZvyšuje dovednost Útok o ${val}.",
+		"vcmi.stackExperience.desc.defense": "{Obrana}\nZvyšuje dovednost Obrana o ${val}.",
+		"vcmi.stackExperience.desc.shots": "{Střely}\nZvyšuje počet střel o ${val}.",
+		"vcmi.stackExperience.desc.casts": "{Seslání}\nZvyšuje počet seslání o ${val}.",
 		"vcmi.stackExperience.table.yes": "Ano",
 	"vcmi.stackExperience.popup.upgradeMultiplier": "Zkušenosti zachované po vylepšení",
 	"vcmi.stackExperience.windowTitle": "Podrobnosti zkušeností oddílu",
