+++ Mods/vcmi/Content/config/translations/english.json
@@ -1038,15 +1038,15 @@
 	"vcmi.stackExperience.table.no": "No",
 	"vcmi.stackExperience.table.shots": "Shots",
 		"vcmi.stackExperience.table.speed": "Speed",
-		"vcmi.stackExperience.desc.experience": "Current stack experience value used to determine rank and all stack experience bonuses.",
-		"vcmi.stackExperience.desc.minDamage": "Minimum damage bonus granted by stack experience rank.",
-		"vcmi.stackExperience.desc.maxDamage": "Maximum damage bonus granted by stack experience rank.",
-		"vcmi.stackExperience.desc.health": "Health bonus granted by stack experience rank.",
-		"vcmi.stackExperience.desc.speed": "Speed bonus granted by stack experience rank.",
-		"vcmi.stackExperience.desc.attack": "Attack bonus granted by stack experience rank.",
-		"vcmi.stackExperience.desc.defense": "Defense bonus granted by stack experience rank.",
-		"vcmi.stackExperience.desc.shots": "Shots bonus granted by stack experience rank.",
-		"vcmi.stackExperience.desc.casts": "Casts bonus granted by stack experience rank.",
+		"vcmi.stackExperience.desc.experience": "{Experience}\nExperience threshold for this rank is ${val}.",
+		"vcmi.stackExperience.desc.minDamage": "{Min Damage}\nIncreases minimum damage by ${val}.",
+		"vcmi.stackExperience.desc.maxDamage": "{Max Damage}\nIncreases maximum damage by ${val}.",
+		"vcmi.stackExperience.desc.health": "{Health}\nIncreases stack health by ${val}.",
+		"vcmi.stackExperience.desc.speed": "{Speed}\nIncreases speed by ${val}.",
+		"vcmi.stackExperience.desc.attack": "{Attack}\nIncreases Attack skill by ${val}.",
+		"vcmi.stackExperience.desc.defense": "{Defense}\nIncreases Defense skill by ${val}.",
+		"vcmi.stackExperience.desc.shots": "{Shots}\nIncreases number of shots by ${val}.",
+		"vcmi.stackExperience.desc.casts": "{Casts}\nIncreases number of casts by ${val}.",
 		"vcmi.stackExperience.table.yes": "Yes",
 	"vcmi.stackExperience.popup.upgradeMultiplier": "Experience kept after Upgrade",
 	"vcmi.stackExperience.windowTitle": "Stack Experience Details",
