+++ client/render/AssetGenerator.h
@@ -83,6 +83,8 @@ private:
 	CanvasPtr createGateListColored(PlayerColor color, PlayerColor backColor) const;
 	CanvasPtr createHeroSlotsColored(PlayerColor backColor) const;
 	CanvasPtr createStackArtifactIndicator(const Point & size) const;
+	CanvasPtr createStackExperienceIcon(const std::string & iconId) const;
+	CanvasPtr createStackExperienceInactiveOverlay() const;
 
 	void createPaletteShiftedSprites();
 	void generatePaletteShiftedAnimation(const AnimationPath & source, const std::vector<PaletteAnimation> & animation);
